+++ src/pages/LetterWritePage/components/index.ts
@@ -4,11 +4,11 @@ import GenderSelect from './inputs/GenderSelect';
 import LetterContent from './inputs/LetterContent';
 import ImageSelect from './inputs/ImageSelect';
 import LetterPaper from './LetterPaper';
-import PolaroidImage from './PolaroidImage';
+import PolaroidImage from './LetterPaper/PolaroidImage';
 import LetterReceiverBottomSheet from './LetterReceiverBottomSheet';
-import LetterReceiverContainer from './LetterReceiverContainer';
+import LetterReceiverContainer from './LetterPaper/LetterReceiverContainer';
 import LetterWriteBottom from './LetterWriteBottom';
-import LetterWriteContent from './LetterWriteContent';
+import LetterWriteContent from './LetterPaper/LetterWriteContent';
 
 export {
   AgeSlider,
